consolidate interpretation tests
